[bgw-gui] Fixed Camera pane drag and drop (#388)
* Set parent on target of camera pane.

* Fixed Drag&Drop for CameraPane.

* Fixed Codestyle issues.
